Refactored into Android library project with intention of pushing to maven central, so it can be a simple one line dependency.
