+++ .gitignore
@@ -1,0 +1,7 @@
+.gradle
+/local.properties
+/.idea/workspace.xml
+/.idea/libraries
+.idea/*
+.DS_Store
+/build
