+++ aes-crypto/gradle.properties
@@ -1,0 +1,4 @@
+# Module Gradle settings.
+POM_NAME=AES-Crypto
+POM_ARTIFACT_ID=aes-crypto
+POM_PACKAGING=jar
