+++ aes-crypto/src/main/AndroidManifest.xml
@@ -1,0 +1,4 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+          package="com.tozny.crypto.android">
+    <application />
+</manifest>
