+++ gradle.properties
@@ -1,0 +1,38 @@
+# Project-wide Gradle settings.
+
+# IDE (e.g. Android Studio) users:
+# Gradle settings configured through the IDE *will override*
+# any settings specified in this file.
+
+# For more details on how to configure your build environment visit
+# http://www.gradle.org/docs/current/userguide/build_environment.html
+
+# Specifies the JVM arguments used for the daemon process.
+# The setting is particularly useful for tweaking memory settings.
+# Default value: -Xmx10248m -XX:MaxPermSize=256m
+# org.gradle.jvmargs=-Xmx2048m -XX:MaxPermSize=512m -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8
+
+# When configured, Gradle will run in incubating parallel mode.
+# This option should only be used with decoupled projects. More details, visit
+# http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
+# org.gradle.parallel=true
+
+VERSION_NAME=0.0.1-SNAPSHOT
+#VERSION_NAME=0.0.1
+VERSION_CODE=1
+GROUP=com.scottyab
+
+
+POM_DESCRIPTION=A simple Android class for encrypting & decrypting strings, aiming to avoid the classic mistakes that most such classes suffer from.
+POM_URL=https://github.com/scottyab/java-aes-crypto.git
+POM_SCM_URL=https://github.com/scottyab/java-aes-crypto.git
+POM_SCM_CONNECTION=scm:git@github.com:scottyab/java-aes-crypto.git
+POM_SCM_DEV_CONNECTION=scm:git@github.com:scottyab/java-aes-crypto.git
+POM_LICENCE_NAME=MIT license
+POM_LICENCE_URL=https://github.com/scottyab/java-aes-crypto/blob/master/LICENSE
+POM_LICENCE_DIST=repo
+POM_DEVELOPER_ID=scottyab
+POM_DEVELOPER_NAME=Scott Alexander-Bown
+
+RELEASE_REPOSITORY_URL=https://oss.sonatype.org/service/local/staging/deploy/maven2
+SNAPSHOT_REPOSITORY_URL=https://oss.sonatype.org/content/repositories/snapshots
