+++ sample/src/main/res/values/strings.xml
@@ -1,0 +1,6 @@
+<resources>
+    <string name="app_name">AESCryptoSample</string>
+
+    <string name="hello_world">AES-Crypto Sample, Hello world!</string>
+    <string name="action_settings">Settings</string>
+</resources>
